Adds protoc-gen-connect-openapi as plugin for apiv2 (#185)

+++ apiv2/test/client/utils_test.go
@@ -14,9 +14,9 @@ import (
 )
 
 // ListAllInstances retrieves all InstanceResource objects by iterating over paginated responses.
-func ListAllInstances(ctx context.Context, client *api.ClientWithResponses, pageSize int32) ([]api.InstanceResource, error) {
+func ListAllInstances(ctx context.Context, client *api.ClientWithResponses, pageSize int) ([]api.InstanceResource, error) {
 	var allInstances []api.InstanceResource
-	var offset int32 = 0
+	var offset int = 0
 
 	for {
 
@@ -44,16 +44,16 @@ func ListAllInstances(ctx context.Context, client *api.ClientWithResponses, page
 		}
 
 		// Increment the offset for the next page
-		offset += int32(pageSize)
+		offset += pageSize
 	}
 
 	return allInstances, nil
 }
 
 // ListAllHosts retrieves all HostResource objects by iterating over paginated responses.
-func ListAllHosts(ctx context.Context, client *api.ClientWithResponses, pageSize int32) ([]api.HostResource, error) {
+func ListAllHosts(ctx context.Context, client *api.ClientWithResponses, pageSize int) ([]api.HostResource, error) {
 	var allHosts []api.HostResource
-	var offset int32 = 0
+	var offset int = 0
 
 	for {
 		// Call the API to get a paginated list of hosts
@@ -87,9 +87,9 @@ func ListAllHosts(ctx context.Context, client *api.ClientWithResponses, pageSize
 }
 
 // ListAllRegions retrieves all RegionResource objects by iterating over paginated responses.
-func ListAllRegions(ctx context.Context, client *api.ClientWithResponses, pageSize int32) ([]api.RegionResource, error) {
+func ListAllRegions(ctx context.Context, client *api.ClientWithResponses, pageSize int) ([]api.RegionResource, error) {
 	var allRegions []api.RegionResource
-	var offset int32 = 0
+	var offset int = 0
 
 	for {
 		// Call the API to get a paginated list of regions
@@ -123,9 +123,9 @@ func ListAllRegions(ctx context.Context, client *api.ClientWithResponses, pageSi
 }
 
 // ListAllSites retrieves all SiteResource objects by iterating over paginated responses.
-func ListAllSites(ctx context.Context, client *api.ClientWithResponses, pageSize int32) ([]api.SiteResource, error) {
+func ListAllSites(ctx context.Context, client *api.ClientWithResponses, pageSize int) ([]api.SiteResource, error) {
 	var allSites []api.SiteResource
-	var offset int32 = 0
+	var offset int = 0
 
 	for {
 		// Call the API to get a paginated list of sites
@@ -159,9 +159,9 @@ func ListAllSites(ctx context.Context, client *api.ClientWithResponses, pageSize
 }
 
 // ListAllWorkloads retrieves all WorkloadResource objects by iterating over paginated responses.
-func ListAllWorkloads(ctx context.Context, client *api.ClientWithResponses, pageSize int32) ([]api.WorkloadResource, error) {
+func ListAllWorkloads(ctx context.Context, client *api.ClientWithResponses, pageSize int) ([]api.WorkloadResource, error) {
 	var allWorkloads []api.WorkloadResource
-	var offset int32 = 0
+	var offset int = 0
 
 	for {
 		// Call the API to get a paginated list of workloads
@@ -202,7 +202,7 @@ func TestDeleteAllRegions(t *testing.T) {
 	apiClient, err := GetAPIClient()
 	require.NoError(t, err)
 
-	var pageSize int32 = 10
+	var pageSize int = 10
 	regions, err := ListAllRegions(ctx, apiClient, pageSize)
 	if err != nil {
 		t.Fatalf("failed to list all regions: %v", err)
@@ -223,7 +223,7 @@ func TestDeleteAllSites(t *testing.T) {
 	apiClient, err := GetAPIClient()
 	require.NoError(t, err)
 
-	var pageSize int32 = 10
+	var pageSize int = 10
 	sites, err := ListAllSites(ctx, apiClient, pageSize)
 	if err != nil {
 		t.Fatalf("failed to list all sites: %v", err)
@@ -244,7 +244,7 @@ func TestDeleteAllWorkloads(t *testing.T) {
 	apiClient, err := GetAPIClient()
 	require.NoError(t, err)
 
-	var pageSize int32 = 10
+	var pageSize int = 10
 	workloads, err := ListAllWorkloads(ctx, apiClient, pageSize)
 	if err != nil {
 		t.Fatalf("failed to list all workloads: %v", err)
@@ -265,7 +265,7 @@ func TestDeleteAllInstances(t *testing.T) {
 	apiClient, err := GetAPIClient()
 	require.NoError(t, err)
 
-	var pageSize int32 = 10
+	var pageSize int = 10
 	instances, err := ListAllInstances(ctx, apiClient, pageSize)
 	if err != nil {
 		t.Fatalf("failed to list all instances: %v", err)
@@ -286,7 +286,7 @@ func TestDeleteAllHosts(t *testing.T) {
 	apiClient, err := GetAPIClient()
 	require.NoError(t, err)
 
-	var pageSize int32 = 10
+	var pageSize int = 10
 	hosts, err := ListAllHosts(ctx, apiClient, pageSize)
 	if err != nil {
 		t.Fatalf("failed to list all hosts: %v", err)
@@ -301,8 +301,8 @@ func TestDeleteAllHosts(t *testing.T) {
 
 func ListAllLocalAccounts(t *testing.T, ctx context.Context, apiClient *api.ClientWithResponses) []api.LocalAccountResource {
 	var allAccounts []api.LocalAccountResource
-	var offset int32
-	var pageSize int32 = 100 // Adjust page size as needed
+	var offset int
+	var pageSize int = 100 // Adjust page size as needed
 
 	for {
 		resList, err := apiClient.LocalAccountServiceListLocalAccountsWithResponse(
